cvFinal/src/requirement.txt
numpy
opencv-python
opencv-contrib-python
Pillow
scipy
tdqm